Introduce automated testing facility
The btrfs-backup program comes with a large set of tests that verify
not only that the program behaves as expected but also that it works in
conjunction with the installed version of sys-fs/btrfs-progs. As such,
it makes sense to run the tests on every installation. Portage supports
testing of the package to install by means of the 'test' feature which,
unfortunately, cannot be enabled by default by the ebuild to emerge
itself.
This change does two things: first, it introduces the functionality to
exercise the tests. It turned out to be impossible to combine setup.py's
test functionality with portage's sandboxing feature: whenever the test
is about to be executed it would download the argparse package (as it is
listed as a dependency) and unpack it locally. Unfortunately, there is
no way to convince the program to change this behavior or to install the
files in a different directory. To that end, the way of invoking Python
with the unit test module pointed to the correct directory had to be
used.
Second, the change adds a warning that is printed if a set of portage
features is in use that prohibits execution of the tests. Unfortunately,
the conditions for the tests to run successful are rather strict. Since
they make use of loop back devices and the mount program, root
privileges are required for the tests to run correctly. That means that
portage must not drop root privileges which can only be achieved with a
certain set of portage features (which are enabled by default) disabled.