Add 'tests' USE flag to allow excluding tests from installation
By default btrfs-backup is installed along with the unit tests verifying
its functionality. In certain cases the user might not want to have
these tests installed, when he/she is unable to run them in the first
place, for example.
This change introduces a new use flag 'tests' (set by default) which can
be used to control whether or not the unit tests should be installed.
Note that unsetting this use flag implies that the automated tests
recommended to run during installation (by means of the portage's test
feature) cannot be run -- this constraint is enforced by the ebuilded.